[WIP] Refactored the Terraform code out of machine actuator
Refactored the Terraform code out of the machine actuator to make it easier to transition
the actuator to govmomi.

Resolves #30
