+++ cloud/vsphere/provisioner/govmomi/delete.go
@@ -1,0 +1,17 @@
+package govmomi
+
+import (
+	clusterv1 "sigs.k8s.io/cluster-api/pkg/apis/cluster/v1alpha1"
+)
+
+type Deleter struct {
+}
+
+func NewDeleter() *Creator {
+	return &Creator{}
+}
+
+// Create the machine.
+func (c *Deleter) Deleter(cluster *clusterv1.Cluster, machine *clusterv1.Machine) error {
+	return nil
+}
